+++ src/libs/Detach/Collisions/Primitives2D/Circle.cs
@@ -4,13 +4,12 @@ namespace Detach.Collisions.Primitives2D;
 
 public record struct Circle
 {
-	// TODO: Rename to Center.
-	public Vector2 Position;
+	public Vector2 Center;
 	public float Radius;
 
-	public Circle(Vector2 position, float radius)
+	public Circle(Vector2 center, float radius)
 	{
-		Position = position;
+		Center = center;
 		Radius = radius;
 	}
 }