+++ src/libs/Detach/Collisions/Primitives3D/Aabb.cs
@@ -4,20 +4,19 @@ namespace Detach.Collisions.Primitives3D;
 
 public record struct Aabb
 {
-	// TODO: Rename to Center.
-	public Vector3 Origin;
+	public Vector3 Center;
 	public Vector3 Size;
 
-	public Aabb(Vector3 origin, Vector3 size)
+	public Aabb(Vector3 center, Vector3 size)
 	{
-		Origin = origin;
+		Center = center;
 		Size = size;
 	}
 
 	public Vector3 GetMin()
 	{
-		Vector3 p1 = Origin + Size / 2;
-		Vector3 p2 = Origin - Size / 2;
+		Vector3 p1 = Center + Size / 2;
+		Vector3 p2 = Center - Size / 2;
 
 		return new Vector3(
 			MathF.Min(p1.X, p2.X),
@@ -27,8 +26,8 @@ public record struct Aabb
 
 	public Vector3 GetMax()
 	{
-		Vector3 p1 = Origin + Size / 2;
-		Vector3 p2 = Origin - Size / 2;
+		Vector3 p1 = Center + Size / 2;
+		Vector3 p2 = Center - Size / 2;
 
 		return new Vector3(
 			MathF.Max(p1.X, p2.X),