+++ src/libs/Detach/Collisions/Primitives3D/Cylinder.cs
@@ -4,14 +4,13 @@ namespace Detach.Collisions.Primitives3D;
 
 public record struct Cylinder
 {
-	// TODO: Rename to BaseCenter.
-	public Vector3 BasePosition;
+	public Vector3 BottomCenter;
 	public float Radius;
 	public float Height;
 
-	public Cylinder(Vector3 basePosition, float radius, float height)
+	public Cylinder(Vector3 bottomCenter, float radius, float height)
 	{
-		BasePosition = basePosition;
+		BottomCenter = bottomCenter;
 		Radius = radius;
 		Height = height;
 	}